+++ package.json
@@ -12,7 +12,7 @@
   "dependencies": {
     "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.11.0",
-    "@mui/icons-material": "^5.14.0",
+    "@mui/icons-material": "^5.14.1",
     "@mui/material": "^5.14.0",
     "@types/echarts": "^4.9.18",
     "axios": "^1.4.0",
