+++ src/components/client/ClientInfoCard.tsx
@@ -8,12 +8,13 @@ import request, { RequestSuccessHandler } from "../../lib/request.ts";
 import { ClientData } from "../../mock/client/clientTable.ts";
 import clientIdState from "../../states/client/ClientIdState.tsx";
 import Button from "@mui/material/Button";
+import { LocationOn, PhoneIphone, Email } from "@mui/icons-material";
 import ClientRemovePopUpButton from "./ClientRemovePopUpButton.tsx";
+import { SvgIcon } from "@mui/material";
 
 type Props = {
   width?: string | number;
   height?: string | number;
-  flexGrow?: number;
 };
 
 function ClientInfoCard({ width, height }: Props) {
@@ -48,14 +49,7 @@ function ClientInfoCard({ width, height }: Props) {
   return (
     <Card sx={{ width, height }}>
       <CardContent>
-        <Box sx={{ display: "flex", justifyContent: "space-between" }}>
-          <Typography
-            sx={{ display: "inline-block" }}
-            variant="h5"
-            contentEditable={editMode}
-          >
-            {name}
-          </Typography>
+        <Box sx={{ display: "flex", justifyContent: "flex-end" }}>
           <Box sx={{ display: "flex", justifyContent: "right", gap: 1 }}>
             {editMode ? (
               <Button
@@ -81,30 +75,47 @@ function ClientInfoCard({ width, height }: Props) {
             <ClientRemovePopUpButton />
           </Box>
         </Box>
+        <br />
         <Typography
-          sx={{ display: "inline-block", fontSize: 14 }}
+          sx={{ display: "inline-block" }}
+          variant="h4"
+          contentEditable={editMode}
+        >
+          {name}
+        </Typography>
+        <br />
+        <hr />
+        <br />
+        <Typography
+          sx={{ display: "inline-block", fontSize: 20 }}
           color="text.secondary"
           gutterBottom
           contentEditable={editMode}
         >
+          <br />
+          <SvgIcon component={Email} /> &nbsp;
           {email}
         </Typography>
         <br />
         <Typography
-          sx={{ display: "inline-block", fontSize: 14 }}
+          sx={{ display: "inline-block", fontSize: 20 }}
           color="text.secondary"
           gutterBottom
           contentEditable={editMode}
         >
+          <br />
+          <SvgIcon component={PhoneIphone} /> &nbsp;
           {phone}
         </Typography>
         <br />
         <Typography
-          sx={{ display: "inline-block", fontSize: 14 }}
+          sx={{ display: "inline-block", fontSize: 20 }}
           color="text.secondary"
           gutterBottom
           contentEditable={editMode}
         >
+          <br />
+          <SvgIcon component={LocationOn} /> &nbsp;
           {address}
         </Typography>
       </CardContent>
