+++ src/hook/useWindowSize.tsx
@@ -1,0 +1,27 @@
+import { useEffect, useState } from "react";
+
+function useWindowSize(): [number, number] {
+  const [windowSize, setWindowSize] = useState({
+    height: window.innerHeight,
+    width: window.innerWidth,
+  });
+
+  const handleResize = () => {
+    setWindowSize({
+      width: window.innerWidth,
+      height: window.innerHeight,
+    });
+  };
+
+  useEffect(() => {
+    window.addEventListener("resize", handleResize);
+
+    return () => {
+      window.removeEventListener("resize", handleResize);
+    };
+  }, []);
+
+  return [windowSize.width, windowSize.height];
+}
+
+export default useWindowSize;
